service4/requirements.txt
Jinja2-2.10.3
MarkupSafe-1.1.1
Werkzeug-0.16.0
click-7.0
Flask-1.1.1
itsdangerous-1.1.0
flask-sqlalchemy-2.4.1
flask-login-0.4.1
pymysql-0.9.3

